Output fasta of features
A new option is added for an output file that will print the
identified rRNA features. To add this functionality additional code
was required to save the results of nhmmer as Bio::SeqFeature::Generic
objects so that they can be properly printed.

This change has required the additional dependancy of some Bioperl
modules:
Bio::SeqFeature::Generic
Bio::Tools::GFF
Bio::SeqIO